restructure nav component
